requirements.txt


Flask==3.1.2
flask-cors==5.0.0
gunicorn==23.0.0



mediapipe==0.10.14
opencv-python-headless==4.10.0.84
numpy==1.26.4
protobuf==4.25.3
pandas==2.2.2
google-generativeai==0.8.5


python-dotenv==1.0.1
Pillow==10.3.0
requests==2.32.5


annotated-types==0.7.0
blinker==1.9.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
google-ai-generativelanguage==0.6.15
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
proto-plus==1.23.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
rsa==4.9
six==1.16.0
tqdm==4.66.4
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.1.3